Rename `Edisi` Screen Ui model
