+++ ui/src/main/java/org/redaksi/ui/edisi/EdisiViewModel.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.redaksi.data.remote.PillarApi
-import org.redaksi.ui.model.IssueWithArticle
+import org.redaksi.ui.model.IssueUi
 import org.redaksi.ui.model.fromResponse
 import javax.inject.Inject
 
@@ -40,6 +40,6 @@ class EdisiViewModel @Inject constructor(private val pillarApi: PillarApi) : Vie
 }
 
 data class EdisiViewModelState(
-    val issuesUi: List<IssueWithArticle> = listOf(),
+    val issuesUi: List<IssueUi> = listOf(),
     val isLoading: Boolean = true
 )
